soil-web-gui/backend/requirements.txt
fastapi
uvicorn
python-multipart
numpy
pillow
matplotlib
